Recipes and Linted
Return 3 recipes given calories and disable specific parts of pylint that are deemed unreasonable
